requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
Werkzeug==3.0.6
Pillow==10.4.0
stripe==11.0.0
email-validator==2.2.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
cloudinary==1.36.0